coffee_shop/requirements-dev.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
crispy-tailwind==1.0.3
decorator==5.2.1
dj-database-url==2.3.0
Django==5.2
django-crispy-forms==2.4
django-environ==0.12.0
djangorestframework==3.16.0
executing==2.2.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
